ミニマップの実装、描画エンジンの仕様変更。

+++ DirectX 初ゲーム/FadeScreen.cpp
@@ -4,7 +4,7 @@ FadeScreen::FadeScreen(void)
 {
 	this->setPolygon(Layer::TOP, Texture::none, RendererType::UI);
 
-	this->polygon->setSize(SystemParameters::ResolutionX, SystemParameters::ResolutionY);
+	this->polygon->setSize((float)SystemParameters::ResolutionX, (float)SystemParameters::ResolutionY);
 	this->polygon->setOpacity(0.0f);
 
 	state = Stop;
